Time: 2 ms (51.77%), Space: 8.4 MB (17.91%) - LeetHub

+++ 0069-sqrtx/0069-sqrtx.cpp
@@ -9,7 +9,7 @@ public:
             mid = s + (e-s)/2;
             if((long)mid*mid == (long)x){
                 return mid;
-            }else if(mid*mid < x){
+            }else if((long)mid*mid < (long)x){
                 s = mid + 1;
             }else{
                 e = mid - 1;
